Replace loops with iterators and chain arc vectors
Rather than manually looping and indexing, we can combine an iterator
with a map(). This has the advantage of removing mutable variables,
and re-allocations on push(), allowing the compiler to elide bounds checks (potentially giving
speedups), and being a bit less error prone.

We can combine the above with the chain() method when building arcs,
which allows us to avoid allocating two vectors and then combining
them into a third; it's possible that the compiler can optimise this
away, but with this approach we don't need to take the chance: using
chain() allows us to make a single allocation.

We're also leveraging Rust's statement-oriented nature here, allowing
constructions like:

let foo = if { //statement } else { //alternate statement };

again allowing us to avoid defining foo twice.